+++ CHANGELOG.md
@@ -10,6 +10,10 @@
 
 ## 22.15
 
+The following resource has been added:
+
+  * POST `/1.1/users/<user_id>/voicemails`
+
 * PUT on `/users/<user_id>/groups` now accepts the group `uuid` as well as the group `id` in its body
 * The `/tenants` endpoints were modified to remove the field `webrtc_video_sip_template_uuid`.
 * `/lines` endpoint now create the SIP, SCCP or custom endpoint if it is specified in the body and associates it to the line.